krishirakshak-ml/training/requirements-train.txt
torch
ultralytics
timm
mlflow
pyyaml
scikit-learn
numpy
pillow
opencv-python-headless

